feat(tests): ユーザーディレクトリ取得関数と古いエンジンディレクトリ作成関数を追加
